Add several new build targets to run clang-format and clang-tidy.
The new build targets and how they work are described in the README.md file.
Basically they allow code to be formated, using clang-format, to reflect
our style guide, which is basically the same as the FairRoot and ROOT ones.
In addition clang-tidy and cpplint check for questionable code constructs.