Allow field maps to work as local B fields in volumes.
The TVirtualMagField::Field(position, B) function requires the position
to be in global co-ordinates. This means that by default B fields are not
in local co-ordinates. ShipFieldMaker has been updated to obtain the
global translation and (Euler) rotations for a given volume and use this to
find the global-to-local transformation for the position so that we can
find the local field. Here, all volumes that use the same field map
distribution have ShipBFieldMap copies that reuse the map data but only
change the global-to-local TGeoMatrix transformation information.
Other changes are made to ensure that copied field maps work as expected.
It is also possible to rescale a local field map for a given individual
volume without the need to have separate map data. All new options are
described in the README.md file.