Bug fixes for using VMC B fields. (#90)
The VMC only correctly sets the magnetic fields for Geant4 when they are defined
in the C++ code where the geometry is defined. This means that additional
fields set using the addVMCFields() function python/geomGeant4.py are only
recognised by ROOT and not Geant4.

To correct this, we need to reuse the ConstructSDandField() function in TG4GeometryManager
so that fields defined in ShipFieldMaker are recognised by both ROOT and Geant4.
This requires that ShipFieldMaker inherits from TG4VUserPostDetConstruction(), where
we overload the virtual Construct() function to make the fields and assign them to volumes;
this effectively renames the earlier ShipFieldMaker::makeFields(inputFile) function.

Corrected bin calculations in ShipBFieldMap that can occur for certain binning choices.

Added field/convertMap.py as a general script for converting a field map text file
into ROOT format for FairShip use.