Add field definition functions to ShipFieldMaker.
It is now possible to define volume B fields using "defineX()"
function calls with the ShipFieldMaker object in python/geomGeant4.py
in addition to reading in the configuration file. The addVMCFields()
function in geomGeant4.py has been changed to pass the ship_geo
object, which is now used to get the z position for the spectrometer
field map definition. The field/README file has been updated to give
details about the new features, as well as being clearer about various
keywords/parameter definitions and options.

There have been some internal changes to ShipFieldMaker:

1) Changed the ShipFieldMaker constructor to only require the verbosity
boolean (default = false). The configuration input file is only processed
when the new readInputFile(configFileName) function is called. This allows
us the option of not requiring a configuration input file to setup fields.

2) Renamed the createX() functions to defineX(), with public versions
for calling via the python scripts (& C++ code) and protected versions
that are used when each line in the configuration input file is processed.
The protected functions simply extract the keywords and parameters from
the input line and pass this information onto the public functions that
do the actual work.

3) The local and regional (local+global) field names that are needed
for each volume are first stored as internal fieldInfo structure
(volumeName, fieldName, scalingFactor) objects in two separate vectors.
The Construct() function, called by the VMC at the end of initialisation,
then loops over these two vectors and sets the fields for each volume.
Before, Construct() could only process the configuration input file to set
the fields. Now, we can set fields using both the config file and/or
function calls.