Fixing total y dim computation (#109)
The total distance should not be the distance between bricks multiplied by the number of rows. Instead, the distance should be multiplied by the number of rows minus one. Otherwise the middle brick is not symmetric with respect to y, and a space between the last brick and the top of the column is observed.