Fix unused variable in MuDISGenerator
First argument of SetPositions was unused and as a result silently
ignored. As we are already changing it, remove default parameters,
which are probably outdated, and could lead to surprising behaviour.
Code using the old signature should break noticeably, and should be
trivial to fix by removing the first argument.