feat(SND): Removing old SND design configurations
The SND detector has too many nuTauTargetDesign options, most of them before the ECN3 setup and currently broken.

Keeping them in the code is only confusing, and it makes harder to update the design without breaking anything.

Therefore, I have removed all the old configurations, keeping only the two most recent:

    Configuration 3, used in the 2018 CDS, with the magnetized target and the possibility of using a field map (disabled by default)

    Configuration 4, currently the default in FairShip, not magnetized target followed by a spectrometer.

I have chosen to keep these two, so we still have one with magnetized target and one with not magnetized target, while we complete the integration of the SND in the muon shield.

This from a logical point of view. From a code point of view, the changes boil basically to:

- Remove all the lines within if (Design == 0), 1 or 2;
- Change the logic to simplify it and making it more clear (avoid if (Design != 4) calls, but strictly require (if Design == 3)). The first option makes code easy to miss when you add a new configuration.